backend-django-docker/requirements.txt
asgiref==3.6.0
coverage==7.2.3
dj-database-url==1.3.0
Django==4.1.7
django-cors-headers==3.14.0
django-environ==0.10.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
psycopg2-binary==2.9.5
PyJWT==2.6.0
pytz==2022.7.1
sqlparse==0.4.3
typing_extensions==4.5.0
tzdata==2022.7
